submission/requirements.txt
matplotlib==3.9.4
numpy==2.1.3
opencv_python_headless==4.7.0.72
pandas==2.2.3
Pillow==11.1.0
python-dotenv==1.1.0
PyYAML==6.0.2
scikit_learn==1.6.1
seaborn==0.13.2
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
scipy==1.15.2
tqdm==4.66.5
